requirements.txt
cbor2==5.2.0
cryptography==3.3.2
pycose==1.0.1
pycryptodome==3.9.9
pyOpenSSL==19.1.0

click==7.*

pydantic>=1.0.0,<2.0.0
fastapi>=0.61.1,<0.62.0
requests==2.28.2
aiofiles
uvicorn>=0.11.6,<0.12.0
h11==0.9.0
httpcore<=0.15.0

pandas==1.0.5
spacy==2.3.0
